% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covsim.R
\name{covsim}
\alias{covsim}
\title{Generation of covariance matrices.}
\usage{
covsim(d)
}
\arguments{
\item{d}{a numeric vector of eigenvalues.}
}
\value{
a covariance matrix.
}
\description{
The function covsim generates a covariance matrix from a given set of eigenvalues.
}
\details{
The function uses a given set of eigenvalues(\eqn{\lambda}), preferably obatained from
\code{BSeigenval} or \code{MODeigenval}, and a random eigenvector structure (\eqn{V})  to
build a covariance matrix as \eqn{V\lambda V'}. The eigenvectors are obtained from a vector of
random uniform numbers between -1 and 1 and using Gram-Schmidt orthogonalization, from the
\code{orthonormalization} function of \code{far} package.
}
\examples{
A <- BSeigenval(INT = 50, dim = 8, sigma = 0.1)
covsim(A)
}
\author{
Santiago Benitez-Vieyra
}

