% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fullgivens.R
\name{fullgivens}
\alias{fullgivens}
\title{Apply a serie of Givens rotation to a covariance matrix. It allows to trasform a covariance
matrix in a correlation matrix without changing the eignevalues.}
\usage{
fullgivens(M)
}
\arguments{
\item{M}{a matrix.}
}
\value{
a correlation matrix.
}
\description{
\code{fullgivens} apply a series of Givens rotation to a matrix in the (i, j) position.
}
\details{
To convert the covariance matrix to a correlation matrix without changing the
eigenvalues, a series of N - 1 Givens rotations (Bendel and Mickey 1978) are needed, following
the algortithm of Davies and Higham (2000). The functions \code{fullgivens} to perform this
serie of rotations. The code for this function is based on the SAS routine detailed in Wicklin
(2013, Appendix C).
}
\examples{
Z <- BSeigenval(INT = 40, dim = 4, sigma = 0.1)
Z
M <- covsim(Z)
N <-fullgivens(M)
eigen(N)$values
}
\author{
Santiago Benitez-Vieyra, based in the SAS routine of Wicklin (2013).
}
\references{
Bendel, R.B. and Mickey, M.R. 1978. Population correlation matrices for sampling
experiments. Communications in Statistics-Simulation and Computation 7:163-182.

Davies, P.I. and Higham, N.J. 2000. Numerically stable generation of correlation
matrices and their factors. BIT 40:640-651.

Wicklin, R. 2013. Simulating data with SAS. Cary, NC:SAS Institute Inc.
}
\seealso{
\code{\link{givensrot}}
}

