% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/multiRV.R
\name{multiRV}
\alias{multiRV}
\title{Multi-set Escoufier's RV coefficient}
\usage{
multiRV(data, vars, ...)
}
\arguments{
\item{data}{A data frame or matrix with two or more numeric variables.}

\item{vars}{A list of vectors indicating the variables to be assigned to each module.}

\item{...}{Additional arguments to be passed to function \code{cov}.}
}
\value{
The multi-set Escoufier's RV coefficient.
}
\description{
multiRV estimates the strength of the association between multiple sets of
variables (modules) as the average of all pair-wise RV coefficients between sets.
}
\details{
This function follows the suggestion of Klingenberg (2009) to estimate
the strength of association among multiple sets of variables (modules), as the average
among all pair-wise Escoufier's RV coefficients between sets. As the RV coeffcient, the
multi-set RV takes values between zero, when modules are completely separated, and one,
when there are no modular structure.
}
\examples{
multiRV(sorbus, vars = list(c(1,2), c(3,4), c(5:7)))
}
\author{
Santiago Benitez-Vieyra
}
\references{
Klingenberg, C.P. 2009. Morphometric integration and modularity in configurations of
landmarks: tools for evaluating a priori hypotheses. Evolution and Development
11:405-421.
}
\seealso{
\code{\link{pairRV}}
}

