% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/pRV.R
\name{pRV}
\alias{pRV}
\title{Pseudovalues of Escoufier's RV coefficient}
\usage{
pRV(data, vars1, vars2, ...)
}
\arguments{
\item{data}{A data frame or matrix with two or more numeric variables}

\item{vars1}{Vector indicating the variables to be assigned to module 1.}

\item{vars2}{Vector indicating the variables to be assigned to module 2.}

\item{...}{Additional arguments to be passed to function \code{cov}.}
}
\value{
the vector of individual jackknife pseudovalues of the RV coeffcient.
}
\description{
pRV estimates the individual jackknife pseudovalues of the Escoufier's RV
coefficient between two sets of variables
}
\details{
The RV coefficient (Escoufier 1973, Robert and Escoufier 1978) is
  first estimated using the complete sample of individuals. Then, one
  individual is ignored in turn, and the same statistic is computed. Finally,
  the pseudovalues are estimated according to the usual jackknife procedure
  (Tuckey 1958, Manly 1997). The pseudovalue of the i individual is defined
  as

  \deqn{pRV_{i} = (N)RV - (N-1)RV_{-i}}{pRV_i = N*RV - (N-1)*RV_-i}

  where N is the number of individuals, and -i indicates that the i
  individual was removed. before estimating INT. The individual pRV values
  allow to estimate change in morphological modularity due to phenotypic
  selection, avoiding the calculation of the change in the full
  variance-covariance phenotypic matrix, following Benitez-Vieyra et al. (in
  prep).
}
\examples{
pRV(sorbus, vars1 = c(3, 4), vars2 = c(1, 2, 5:7))
}
\author{
Santiago Benitez-Vieyra
}
\references{
Benitez-Vieyra, S., Fornoni, J., Dominguez, C.A. in prep.

Escoufier, Y. 1973. Le traitement des variables vectorielles.
Biometrics 29:751-760.

Manly, B.J.S., 1997. Randomization, Bootstrap and Monte Carlo
methods in Biology. CRC Press, Boca Raton.

Robert, P., and Escoufier, Y. 1976. A unifying tool for linear
multivariate statistical analysis: the RV-coefficient. Applied Statistics
25:257-265.

Tukey, J.W. 1958. Bias and confidence in not-quite large samples.
Annals of Mathematical Statistics 29:614.
}
\seealso{
\code{\link{pairRV}}
}

