% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{sorbus}
\alias{sorbus}
\title{Phenotypic and fitness meassures in Sorbus terminalis}
\format{a data set with 27 rows and 9 variables.}
\usage{
data(sorbus)
}
\description{
Data from \emph{Sorbus terminalis} (population Hoyos de Muñoz), including
measures from seven floral traits, plant size and plant fitness. Variables
are PL, petal length; CD, corolla diameter; SL, style length; HK, herkogamy;
HL, hypanthium length; HW, hypanthium width; GW, gynoecium width; BASAR,
basal area (plant size) and W, fitness measured as fruit production.
}
\details{
This dataset was kindly provided by Mariano Ordano and co-workers. Please,
cite as Ordano et al. 2008 (see references).
}
\examples{
data(sorbus)
}
\references{
Ordano, M., Fornoni, J., Boege, K. and Domínguez, C.A. 2008. The
  adaptive value of phenotypic floral integration. New Phytologist
  179:1183–1192.
}

