\name{selint-package}
\alias{selint-package}
\alias{selint}
\docType{package}
\title{
Phenotypic integration in R 
}
\description{
A package to estimate the phenotypic integration index, the RV coeffcient of modularity and their jackknife pseudovalues. Pseudovalues allow testing biological hypotheses about the adaptive significance of phenotypic integration and different patterns of modularity, using simple phenotypic selection approaches.
}

\author{
Code contributed by: Santiago Benitez-Vieyra  

Package maintained by: Santiago Benitez-Vieyra (santiagombv@gmail.com)  

Development version hosted on GitHub: https://github.com/santiagombv/selint.git 

Please file bug reports at: https://github.com/santiagombv/selint/issues  
}
\references{
Benitez-Vieyra S, Fornoni J, Dominguez CA (submitted) A practical guide to measure natural selection on phenotypic integration.
}
