\name{selint-package}
\alias{selint-package}
\alias{selint}
\docType{package}
\title{
Phenotypic integration in R 
}
\description{
A package to estimate the phenotypic integration index, the RV coeffcient of modularity and their jackknife pseudovalues. This allows testing biological hypotheses about the adaptive significance of different patterns of integration, using simple phenotypic selection approaches.
}
\details{
\tabular{ll}{
Package: \tab selint\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-09\cr
License: \tab GPL-2 \cr
}

}
\author{
Code contributed by: Santiago Benitez-Vieyra

Package maintained by: Santiago Benitez-Vieyra (santiagombv@gmail.com
)

Development version hosted on GitHub: https://github.com/santiagombv/selint.git

Please file bug reports at: https://github.com/santiagombv/selint/issues

}
\references{
Benitez-Vieyra S, Fornoni J, Dominguez CA (in prep.) Phenotypic selection on individual integration.
}
